+++ src/main/java/com/chrisrm/ideaddlc/UIReplacer.java
@@ -46,6 +46,8 @@ import com.intellij.ui.Gray;
 import com.intellij.ui.JBColor;
 import com.intellij.ui.tabs.FileColorManagerImpl;
 import com.intellij.ui.tabs.TabsUtil;
+import com.intellij.ui.tabs.UiDecorator;
+import com.intellij.ui.tabs.newImpl.JBTabsImpl;
 import com.intellij.util.ObjectUtils;
 import com.intellij.util.containers.ContainerUtil;
 import com.intellij.util.ui.JBUI;
@@ -72,11 +74,11 @@ public enum UIReplacer {
 
   public static void patchUI() {
     try {
+      patchAutocomplete();
       patchTabs();
       patchTables();
       patchGrays();
       patchMemoryIndicator();
-      patchAutocomplete();
       patchIdeaActionButton();
       patchScrollbars();
       patchDialogs();
@@ -159,41 +161,26 @@ public enum UIReplacer {
   /**
    * Patch the autocomplete color with the accent color
    */
-  static void patchAutocomplete() throws NoSuchFieldException, IllegalAccessException {
+  static void patchAutocomplete() throws IllegalAccessException {
     if (!MTConfig.getInstance().isMaterialTheme()) {
       return;
     }
     final Color defaultValue = UIUtil.getListSelectionBackground();
-    final Color autocompleteSelectionBackground = ObjectUtils.notNull(UIManager.getColor("Autocomplete.selectionBackground"), defaultValue);
-    final Color autocompleteSelectionForeground = ObjectUtils.notNull(UIManager.getColor("Autocomplete.selectionForeground"), defaultValue);
-    final Color autocompleteSelectionForegroundGreyed = ObjectUtils.notNull(UIManager.getColor("Autocomplete.selectionForegroundGreyed"), defaultValue);
-    final Color autoCompleteBackground = ObjectUtils.notNull(UIManager.getColor("Autocomplete.background"), defaultValue);
-    final Color autocompleteForeground = ObjectUtils.notNull(UIManager.getColor("Autocomplete.foreground"), defaultValue);
-    final Color autocompleteSelectionUnfocused = ObjectUtils.notNull(UIManager.getColor("Autocomplete.selectionUnfocus"), defaultValue);
-    final Color autocompleteSelectedGreyedForeground = ObjectUtils.notNull(UIManager.getColor("Autocomplete.selectedGreyedForeground"), defaultValue);
-    final Color autocompletePrefixForegroundColor = ObjectUtils.notNull(UIManager.getColor("Autocomplete.prefixForeground"), defaultValue);
-    final Color autocompleteSelectedPrefixForegroundColor = ObjectUtils.notNull(UIManager.getColor("Autocomplete.selectedPrefixForeground"), defaultValue);
-    final Color autocompleteUserSelectedPrefixForegroundColor = ObjectUtils.notNull(UIManager.getColor("Autocomplete.userSelectedPrefixForeground"), defaultValue);
+    final Color autoCompleteBackground = ObjectUtils.notNull(UIManager.getColor("CompletionPopup.background"), defaultValue);
 
     final Field[] fields = LookupCellRenderer.class.getDeclaredFields();
-    final Object[] colorFields = Arrays.stream(fields)
+    Arrays.stream(fields)
         .filter(f -> f.getType().equals(Color.class))
-        .toArray();
+        .filter(field -> field.getName().equals("BACKGROUND_COLOR"))
+        .findFirst()
+        .ifPresent(field -> {
+          try {
+            StaticPatcher.setFinalStatic(field, autoCompleteBackground);
+          } catch (NoSuchFieldException | IllegalAccessException e) {
+            System.err.println("Unable to hack autocomplete: " + e.getLocalizedMessage());
+          }
+        });
 
-    StaticPatcher.setFinalStatic((Field) colorFields[0], autoCompleteBackground);
-    StaticPatcher.setFinalStatic((Field) colorFields[1], autocompleteForeground);
-    StaticPatcher.setFinalStatic((Field) colorFields[2], autocompleteSelectedGreyedForeground);//grayed fore ground
-    StaticPatcher.setFinalStatic((Field) colorFields[3], autocompleteSelectionBackground);//selected background color
-    StaticPatcher.setFinalStatic((Field) colorFields[4], autocompleteSelectionUnfocused);//selected non focused background color
-    StaticPatcher.setFinalStatic((Field) colorFields[5], autocompleteSelectionForeground);//selected foreground color
-    StaticPatcher.setFinalStatic((Field) colorFields[6], autocompleteSelectionForegroundGreyed);//selected grayed foreground color
-    StaticPatcher.setFinalStatic((Field) colorFields[7], autocompletePrefixForegroundColor);//prefix foreground color
-    StaticPatcher.setFinalStatic((Field) colorFields[8], autocompleteSelectedPrefixForegroundColor);//selected prefix foreground color
-    Optional.of(colorFields)
-        .filter(colorField -> colorField.length > 9)
-        .map(colorField -> (Field) colorFields[9])
-        .ifPresent(colorField -> ToolBoxKt.runSafely(()->
-            StaticPatcher.setFinalStatic(colorField, autocompleteUserSelectedPrefixForegroundColor)));//selected prefix foreground color
   }
 
 
@@ -475,9 +462,13 @@ public enum UIReplacer {
    */
   public static void patchTabs() throws NoSuchFieldException, IllegalAccessException {
     LegacySupportUtility.INSTANCE.invokeClassSafely("com.intellij.ide.ui.laf.darcula.ui.DarculaSeparatorUI", () -> {
-      final int baseHeight = JBUI.scale(6);
-      final int tabsHeight = MTConfig.getInstance().getTabsHeight() / 2 - baseHeight;
+      final int baseHeight = 9;
+      final int tabsHeight = Math.max(MTConfig.getInstance().getTabsHeight() / 2 - baseHeight, 0);
       StaticPatcher.setFinalStatic(TabsUtil.class, "TAB_VERTICAL_PADDING", new JBValue.Float(tabsHeight));
+      StaticPatcher.setFinalStatic(TabsUtil.class, "NEW_TAB_VERTICAL_PADDING", tabsHeight);
+
+      StaticPatcher.setFinalStatic(JBTabsImpl.class, "ourDefaultDecorator",
+          (UiDecorator) () -> new UiDecorator.UiDecoration(null, JBUI.insets(TabsUtil.NEW_TAB_VERTICAL_PADDING, 8)));
     });
   }
 }