+++ src/main/java/io/acari/DDLC/Constants.kt
@@ -8,9 +8,9 @@ const val DARK_MODE_CONSTANT: String = "Ḑ̶͍̲̳̫͗̚͠ä̵͉̠̬͙̤͚͆͆
 val UPDATE_MESSAGE: String = """
       What's New?<br>
       <ul>
-      <li>Doki-Doki Icons do not get overridden by other themes now.</li>
-      <li>More Kotlin icon associations.</li>
-      <li>Less icon residue when switching themes.</li>
+      <li>2019.2 Build Support!</li>
+      <li>Up to date with Material UI 3.10</li>
+      <li>Doki icons are overridden by Material UI, to fix just "Toggle Material Icons"</li>
       <br>
       Thanks again for downloading <b>The Doki-Doki Theme</b>! •‿•<br>
       <br>Please see the <a href="https://github.com/cyclic-reference/ddlc-jetbrains-theme/blob/master/docs/CHANGELOG.md">Changelog</a> for more details.