+++ src/main/java/io/acari/DDLC/hax/DokiExtensionHacker.kt
@@ -2,8 +2,11 @@ package io.acari.DDLC.hax
 
 import com.chrisrm.ideaddlc.MTThemeManager
 import com.intellij.ide.IconProvider
+import com.intellij.openapi.extensions.impl.ExtensionComponentAdapter
 import com.intellij.openapi.extensions.impl.ExtensionPointImpl
+import com.intellij.openapi.util.IconLoader
 import com.intellij.util.containers.ContainerUtil
+import com.intellij.util.containers.ImmutableList
 import com.intellij.util.containers.stream
 import io.acari.DDLC.LegacySupportUtility
 import io.acari.DDLC.legacy.Runner
@@ -13,9 +16,13 @@ import java.util.stream.Collectors
 
 object DokiExtensionHacker {
   private lateinit var cacheFeildToHack: Field
+  private lateinit var adapterFeildToHack: Field
   private var materialExtensions: MutableList<Any?> = ArrayList()
   private var nonMaterialExtensions: MutableList<Any?> = ArrayList()
   private var dokiExtensions: MutableList<Any?> = ArrayList()
+  private var materialAdapters: MutableList<Any?> = ArrayList()
+  private var nonMaterialAdapters: MutableList<Any?> = ArrayList()
+  private var dokiAdapters: MutableList<Any?> = ArrayList()
 
   init{
     MTThemeManager.addMaterialThemeActivatedListener {
@@ -59,6 +66,22 @@ object DokiExtensionHacker {
   private fun mutilateExtensionPoints() {
     val classToHack = ExtensionPointImpl::class.java
     cacheFeildToHack = classToHack.declaredFields.stream().filter { it.name == "myExtensionsCache" }.findFirst().get()
+
+    val (dokiExtractedExtensions, materialExtractedExtensions, nonMaterialExtractedExtensions) =
+        mutilateField(cacheFeildToHack) { it?.javaClass?.name ?: ""}
+    dokiExtensions = dokiExtractedExtensions
+    materialExtensions = materialExtractedExtensions
+    nonMaterialExtensions = nonMaterialExtractedExtensions
+
+    adapterFeildToHack = classToHack.declaredFields.stream().filter { it.name == "myAdapters" }.findFirst().get()
+    val (dokiExtractedAdapters, materialExtractedAdapters, nonMaterialExtractedAdapters) =
+        mutilateField(adapterFeildToHack) { (it as? ExtensionComponentAdapter)?.implementationClass?.name ?: ""}
+    dokiAdapters = dokiExtractedAdapters
+    materialAdapters = materialExtractedAdapters
+    nonMaterialAdapters = nonMaterialExtractedAdapters
+  }
+
+  private fun mutilateField(cacheFeildToHack: Field, nameExtractor: (Any?)->String): Triple<ImmutableList<Any?>, ImmutableList<Any?>, ImmutableList<Any?>> {
     cacheFeildToHack.isAccessible = true
     val point = IconProvider.EXTENSION_POINT_NAME.getPoint(null)
     point.extensionList
@@ -66,23 +89,30 @@ object DokiExtensionHacker {
     if (cache is List<*>) {
       val partitionedIconProviders: Map<String, List<Any?>> = cache.stream().collect(
           Collectors.groupingBy {
-            val name = it?.javaClass?.name ?: ""
+            val name = nameExtractor(it)
             val containsMaterial = name.contains("com.chrisrm")
             val containsDDLC = name.contains("ddlc")
-            if(!containsDDLC && containsMaterial) "MATERIAL"
-            else if(containsDDLC && containsMaterial) "DOKI-DOKI"
+            if (!containsDDLC && containsMaterial) "MATERIAL"
+            else if (containsDDLC && containsMaterial) "DOKI-DOKI"
             else "OTHER"
           }
       )
 
-      this.nonMaterialExtensions = ContainerUtil.immutableList(partitionedIconProviders["OTHER"] ?: Collections.emptyList())
-      this.materialExtensions = ContainerUtil.immutableList(partitionedIconProviders["MATERIAL"] ?: Collections.emptyList())
-      this.dokiExtensions = ContainerUtil.immutableList(partitionedIconProviders["DOKI-DOKI"] ?: Collections.emptyList())
+      val nonMaterialExtensions_2 = ContainerUtil.immutableList(partitionedIconProviders["OTHER"]
+          ?: Collections.emptyList())
+      val materialExtensions_2 = ContainerUtil.immutableList(partitionedIconProviders["MATERIAL"]
+          ?: Collections.emptyList())
+      val dokiExtensions_2 = ContainerUtil.immutableList(partitionedIconProviders["DOKI-DOKI"]
+          ?: Collections.emptyList())
 
-      val arrayList = ArrayList(dokiExtensions)
-      ContainerUtil.addAll(arrayList, nonMaterialExtensions)
+      val arrayList = ArrayList(dokiExtensions_2)
+      ContainerUtil.addAll(arrayList, nonMaterialExtensions_2)
       cacheFeildToHack.set(point, ContainerUtil.immutableList(arrayList))
+      return Triple(dokiExtensions_2,
+          materialExtensions_2,
+          nonMaterialExtensions_2)
     }
-
+    val emptyList = ContainerUtil.immutableList(mutableListOf<Any?>())
+    return Triple(emptyList,emptyList,emptyList)
   }
 }